Add Handler folder in App folder Move Counter.php to Handler folder Edit index.php autoloader.php
[#1.2]
