+++ index.php
@@ -1,9 +1,9 @@
 <?php
 
-use App\Counter;
 use App\FileCsv;
 use App\FileCsvReader;
 use App\Filter;
+use App\Handler\Counter;
 
 require 'autoloader.php';
 const DEFAULT_OUTPUT_FORMAT   = 'csv';
